Hello Bro !   Notes made from your tutorial
Hi andi, I watched your git-hub tutorial and made a notes out of it. Just for practice I am making a pull request to you to add this file. And you made delete it if you wish to. 
I did this pull request for my practice as well...

Once again thank you so much for your tutorials.